# perf: Introduce sort prefix computation for early TopK exit optimization on partially sorted input (10x speedup on top10 bench)