--- a/datafusion/physical-plan/src/sorts/sort.rs
+++ b/datafusion/physical-plan/src/sorts/sort.rs
@@ -1084,7 +1084,7 @@ impl SortExec {
             .equivalence_properties()
             .extract_common_sort_prefix(&requirement);
 
-        let sort_partially_satisfied = sort_satisfied || !sort_prefix.is_empty();
+        let sort_partially_satisfied = !sort_prefix.is_empty();
 
         // The emission type depends on whether the input is already sorted:
         // - If already fully sorted, we can emit results in the same way as the input