… datafusion/core/tests/physical_optimizer/enforce_sorting.rs
… datafusion/sqllogictest/test_files/topk.slt
… datafusion/sqllogictest/test_files/window.slt