@alamb 2025-04-07
🤖 `./gh_compare_branch.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking) Running
Linux aal-dev 6.8.0-1016-gcp #18-Ubuntu SMP Fri Oct  4 22:16:29 UTC 2024 x86_64 x86_64 x86_64 GNU/Linux
Comparing perf/partitioned_topk (9e8968856543d41bdc05a85f2002c4a8494c6b58) to 19a1e58071cd60e74faabab805d97fd269418925 [diff](https://github.com/apache/datafusion/compare/19a1e58071cd60e74faabab805d97fd269418925..9e8968856543d41bdc05a85f2002c4a8494c6b58)
Benchmarks: sort_tpch clickbench_1 tpch_mem
Results will be posted here when complete