@berkaysynnada 2025-04-09
> > @geoffreyclaude I'm going to merge if you're done with this?
> 
> @berkaysynnada: Fine by me! We resolved the (unfortunately, necessary) semi-redundant common prefix field.
> 
> Only remaining question is the Emission type: if it makes sense to leave it as `Both` you can merge, but if you think `Incremental` makes more sense I can do a quick new commit to change it.

Let's make it incremental for consistency. We're thinking on this, and likely figure out a refactor to make all things more clear. (if you wonder, there is also an issue here https://github.com/apache/datafusion/pull/15479#issuecomment-2786362233. These two cases seem to be related and I believe I'll solve them all)