@geoffreyclaude 2025-04-09
> Let's make it incremental for consistency.
@berkaysynnada: Looking at this again, I'm pretty sure it should actually be `Final`! On partially sorted input, with or without fetch, the SortExec will only emit once it has processed all it's necessary input data (eventually cancelling the input stream early). I've updated the PR with a new fix commit.