@berkaysynnada 2025-04-09
> > Let's make it incremental for consistency.
> 
> @berkaysynnada: Looking at this again, I'm pretty sure it should actually be `Final`! On partially sorted input, with or without fetch, the SortExec will only emit once it has processed all its necessary input data (eventually cancelling the input stream early). I've updated the PR with a new fix commit.

PartiallySorted case is actually the same of AggregateExec with an ordered group_by key. Final means the operator can only emit results after its inputs are fully consumed.