@geoffreyclaude 2025-04-09
> PartiallySorted case is actually the same of AggregateExec with an ordered group_by key. 
Exactly, expect for now the SortExec without fetch does *not* incrementally emit results on partially sorted input: it behaves the same as on completely unsorted input. And with fetch (so TopK), it still buffers everything in memory before emitting its results.
I think the important thing here is that all records are emitted only once the node's execution terminates (whether all the input has been read or not should be irrelevant to the emission type.)