@berkaysynnada 2025-04-09
> Exactly, except for now the SortExec without fetch does _not_ incrementally emit results on partially sorted input: it behaves the same as on completely unsorted input. 

Oh, you are right. I was thinking that case as if it runs like PartialSortExec.

> And with fetch (so TopK), it still buffers everything in memory before emitting its results. I think the important thing here is that all records are emitted only once the node's execution terminates (whether all the input has been read or not should be irrelevant to the emission type.)

I'm sorry, I can realize the case just now. You are right about partial sort case is Final indeed in SortExec, whether there is a fetch or not