@berkaysynnada 2025-04-09
> All ready :) I need to look at PartialSortExec, I wasn't aware of it and it seems like there could be some code duplication?

No need to keep this open anymore. If you see any possible improvements, we can work on those in the new PR's. 

Thank you @geoffreyclaude, happy to collaborate with you.